Handle CSV app start failure Handle selector expression
