+++ src/Package/Impl/DataInspect/Office/CsvAppFileIO.cs
@@ -2,15 +2,19 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using System;
+using System.ComponentModel;
 using System.Diagnostics;
+using System.Globalization;
 using System.IO;
 using System.Threading;
 using Microsoft.Common.Core;
+using Microsoft.R.Components.ContentTypes;
 using Microsoft.R.Components.InteractiveWorkflow;
 using Microsoft.VisualStudio.R.Package.Shell;
 using Microsoft.VisualStudio.R.Package.Utilities;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
+using Microsoft.VisualStudio.Utilities;
 using static System.FormattableString;
 using Task = System.Threading.Tasks.Task;
 
@@ -18,7 +22,7 @@ namespace Microsoft.VisualStudio.R.Package.DataInspect.Office {
     internal static class CsvAppFileIO {
         private static int _busy;
 
-        public static async Task OpenDataCsvApp(string variableName, string csvFileName) {
+        public static async Task OpenDataCsvApp(string expression) {
             if (Interlocked.Exchange(ref _busy, 1) > 0) {
                 return;
             }
@@ -33,6 +37,11 @@ namespace Microsoft.VisualStudio.R.Package.DataInspect.Office {
                 Directory.CreateDirectory(folder);
             }
 
+            // Strip "base::.GlobalEnv$"
+            var index = expression.IndexOf(VariableProvider.GlobalEnvironmentExpression);
+            var variableName = index >= 0 ? expression.Substring(index + VariableProvider.GlobalEnvironmentExpression.Length + 1) : expression;
+
+            var csvFileName = MakeCsvFileName(variableName);
             var file = ProjectUtilities.GetUniqueFileName(folder, csvFileName, "csv", appendUnderscore: true);
             var rfile = file.Replace('\\', '/');
 
@@ -48,7 +57,14 @@ namespace Microsoft.VisualStudio.R.Package.DataInspect.Office {
                     await e.RespondAsync(Invariant($"write.csv({variableName}, file='{rfile}')") + Environment.NewLine);
                 }
 
-                Task.Run(() => Process.Start(file)).DoNotWait();
+                Task.Run(() => {
+                    try {
+                        Process.Start(file);
+                    } catch(Win32Exception ex) {
+                        VsAppShell.Current.ShowErrorMessage(
+                            string.Format(CultureInfo.InvariantCulture, Resources.Error_CannotOpenCsv, ex.Message));
+                    }
+                }).DoNotWait();
             } finally {
                 await SetStatusTextAsync(currentStatusText);
             }
@@ -61,6 +77,32 @@ namespace Microsoft.VisualStudio.R.Package.DataInspect.Office {
                 var statusBar = VsAppShell.Current.GetGlobalService<IVsStatusbar>(typeof(SVsStatusbar));
                 statusBar.SetText(text);
             });
+        }
+
+        private static string MakeCsvFileName(string variableName) {
+            var project = ProjectUtilities.GetActiveProject();
+            var projectName = project?.FileName;
+
+            var contentTypeService = VsAppShell.Current.ExportProvider.GetExportedValue<IContentTypeRegistryService>();
+            var viewTracker = VsAppShell.Current.ExportProvider.GetExportedValue<IActiveWpfTextViewTracker>();
+
+            var activeView = viewTracker.GetLastActiveTextView(contentTypeService.GetContentType(RContentTypeDefinition.ContentType));
+            var filePath = activeView.GetFilePath();
+
+            var csvFileName = string.Empty;
+
+            if (!string.IsNullOrEmpty(projectName)) {
+                csvFileName += Path.GetFileNameWithoutExtension(projectName);
+                csvFileName += "_";
+            }
+
+            if (!string.IsNullOrEmpty(filePath)) {
+                csvFileName += Path.GetFileNameWithoutExtension(filePath);
+                csvFileName += "_";
+            }
+
+            csvFileName += variableName.Replace('.', '_').Replace('@', '_').Replace('$', '_');
+            return csvFileName;
         }
     }
 }
